Use email component for user inputted email field
The leavers email field was not using the email component previously
therefore there was no validation on what the user inputted. This has
resulted in some bad email addresses being entered which cause
submission failures.

Swap it out for the email component field which has some validtion on
it. Obviously we cannot check whether the email is actually correct, but
we can at least check it is formatted correctly.